correct fix for JR6_* QuickMagIC volume issues

+++ JR6_txt_magic.py
@@ -126,7 +126,7 @@ def main(command_line=True, **kwargs):
         er_location_name = kwargs.get('er_location_name', '')
         noave = kwargs.get('noave', 0) # default (0) means DO average
         meth_code = kwargs.get('meth_code', "LP-NO")
-        volume = kwargs.get('volume', 0.025*0.025*0.025)
+        volume = float(kwargs.get('volume', 0.025*0.025*0.025))
 
     # format variables
     mag_file = input_dir_path+"/" + mag_file
